csvLLM/requirements.txt
pandas>=1.0.0
google-genai>=0.1.0